+++ NEWS.md
@@ -2,6 +2,8 @@
 
 # development version
 
+- Add Cleary 2023 risdiplam ([doi:10.1007/s40262-023-01241-7](https://doi.org/10.1007/s40262-023-01241-7)) - healthy adults and patients with spinal muscular atrophy aged 2 months to 61 years.
+
 - Add Wu 2023 SPI-62 ([doi:10.1007/s40262-023-01278-8](https://doi.org/10.1007/s40262-023-01278-8)) - healthy adults.
 
 - Disambiguated the overloaded `OC` name, which denoted five unrelated